refactor: consolidate memory integration tests and fix working memory filtering
