+++ packages/core/src/memory/index.ts
@@ -1,3 +1,4 @@
 export * from './types';
 export * from './memory';
 export * from './mock';
+export * from './working-memory-utils';
