+++ packages/core/src/memory/working-memory-utils.test.ts
