+++ packages/core/src/memory/working-memory-utils.ts
