+++ packages/fastembed/src/index.ts
@@ -3,7 +3,7 @@ import os from 'node:os';
 import path from 'node:path';
 import { customProvider as customProviderV2 } from '@internal/ai-sdk-v5';
 import { customProvider as customProviderLegacy } from '@internal/ai-sdk-v4';
-import type { EmbeddingModel as EmbeddingModelV3 } from '@internal/ai-v6';
+import type { EmbeddingModelV3 } from '@internal/ai-v6';
 import { customProvider as customProviderV3 } from '@internal/ai-v6';
 import { FlagEmbedding, EmbeddingModel } from 'fastembed';
