+++ packages/memory/integration-tests/src/working-memory.test.ts
@@ -1,1433 +1,21 @@
-import { randomUUID } from 'node:crypto';
-import { mkdtemp } from 'node:fs/promises';
-import { tmpdir } from 'node:os';
-import { join } from 'node:path';
 import { openai } from '@ai-sdk/openai';
-import type { AgentGenerateOptions, MastraDBMessage } from '@mastra/core/agent';
-import { Agent } from '@mastra/core/agent';
-import { fastembed } from '@mastra/fastembed';
-import { LibSQLVector, LibSQLStore } from '@mastra/libsql';
-import { Memory } from '@mastra/memory';
-import type { ToolCallPart, ToolResultPart, TextPart } from 'ai';
+import { openai as openaiV6 } from '@ai-sdk/openai-v6';
 import { config } from 'dotenv';
-import type { JSONSchema7 } from 'json-schema';
-import { describe, expect, it, beforeEach, afterEach } from 'vitest';
-
-// Helper to extract text content from MastraDBMessage
-function getTextContent(message: any): string {
-  if (typeof message.content === 'string') {
-    return message.content;
-  }
-  if (message.content?.parts && Array.isArray(message.content.parts)) {
-    // Concatenate all text parts
-    const textParts = message.content.parts.filter((p: any) => p.type === 'text' && p.text).map((p: any) => p.text);
-    if (textParts.length > 0) {
-      return textParts.join(' ');
-    }
-  }
-  // Fallback: check if content has a direct text property
-  if (message.content?.text) {
-    return message.content.text;
-  }
-  // Fallback: check if content.content exists (nested structure)
-  if (message.content?.content && typeof message.content.content === 'string') {
-    return message.content.content;
-  }
-  return '';
-}
-import { z } from 'zod';
-
-const resourceId = 'test-resource';
-let messageCounter = 0;
-
-// Test helpers
-const createTestThread = (title: string, metadata = {}) => ({
-  id: randomUUID(),
-  title,
-  resourceId,
-  metadata,
-  createdAt: new Date(),
-  updatedAt: new Date(),
-});
-
-const createTestMessage = (threadId: string, content: string, role: 'user' | 'assistant' = 'user'): MastraDBMessage => {
-  messageCounter++;
-  return {
-    id: randomUUID(),
-    threadId,
-    content: {
-      format: 2,
-      parts: [{ type: 'text', text: content }],
-    },
-    role,
-    createdAt: new Date(Date.now() + messageCounter * 1000),
-    resourceId,
-  };
-};
-
-const createToolCallMessage = (threadId: string, toolName: string, args: Record<string, any> = {}): MastraDBMessage => {
-  messageCounter++;
-  const toolCallId = randomUUID();
-  return {
-    id: randomUUID(),
-    threadId,
-    content: {
-      format: 2,
-      parts: [
-        {
-          type: 'tool-invocation',
-          toolInvocation: {
-            state: 'result',
-            toolCallId,
-            toolName,
-            args,
-            result: {},
-          },
-        },
-      ],
-    },
-    role: 'assistant',
-    createdAt: new Date(Date.now() + messageCounter * 1000),
-    resourceId,
-  };
-};
-
-const createToolResultMessage = (
-  threadId: string,
-  toolName: string,
-  toolCallId: string,
-  result: any,
-  extraParts: (TextPart | ToolCallPart | ToolResultPart)[] = [],
-): MastraDBMessage => {
-  messageCounter++;
-  return {
-    id: randomUUID(),
-    threadId,
-    content: {
-      format: 2,
-      parts: [
-        {
-          type: 'tool-invocation',
-          toolInvocation: {
-            state: 'result',
-            toolCallId,
-            toolName,
-            args: {},
-            result,
-          },
-          ...extraParts,
-        },
-      ],
-    },
-    role: 'assistant',
-    createdAt: new Date(Date.now() + messageCounter * 1000),
-    resourceId,
-  };
-};
-
-function extractUserData(obj: any) {
-  // Remove common schema keys
-  const { type, properties, required, additionalProperties, $schema, ...data } = obj;
-  return data;
-}
+import { getWorkingMemoryTests } from './shared/working-memory';
 
 config({ path: '.env.test' });
 
-describe('Working Memory Tests', () => {
-  let memory: Memory;
-  let thread: any;
-  let storage: LibSQLStore;
-  let vector: LibSQLVector;
-
-  describe('Working Memory Test with Template', () => {
-    beforeEach(async () => {
-      // Create a new unique database file in the temp directory for each test
-      const dbPath = join(await mkdtemp(join(tmpdir(), `memory-working-test-${Date.now()}`)), 'test.db');
-      console.log('dbPath', dbPath);
-
-      storage = new LibSQLStore({
-        id: 'test-storage',
-        url: `file:${dbPath}`,
-      });
-      vector = new LibSQLVector({
-        connectionUrl: `file:${dbPath}`,
-        id: 'test-vector',
-      });
-
-      // Create memory instance with working memory enabled
-      memory = new Memory({
-        options: {
-          workingMemory: {
-            enabled: true,
-            template: `# User Information
-- **First Name**: 
-- **Last Name**: 
-- **Location**: 
-- **Interests**: 
-`,
-          },
-          lastMessages: 10,
-          semanticRecall: {
-            topK: 3,
-            messageRange: 2,
-          },
-          generateTitle: false,
-        },
-        storage,
-        vector,
-        embedder: fastembed,
-      });
-      // Reset message counter
-      messageCounter = 0;
-      // Create a new thread for each test
-      thread = await memory.saveThread({
-        thread: createTestThread('Working Memory Test Thread'),
-      });
-    });
-
-    afterEach(async () => {
-      //@ts-ignore
-      await storage.client.close();
-      //@ts-ignore
-      await vector.turso.close();
-    });
-
-    it('should handle LLM responses with working memory using OpenAI (test that the working memory prompt works)', async () => {
-      const agent = new Agent({
-        id: 'memory-test-agent',
-        name: 'Memory Test Agent',
-        instructions: 'You are a helpful AI agent. Always add working memory tags to remember user information.',
-        model: openai('gpt-4o'),
-        memory,
-      });
-
-      await agent.generateLegacy('Hi, my name is Tyler and I live in San Francisco', {
-        threadId: thread.id,
-        resourceId,
-      });
-
-      // Get working memory
-      const workingMemory = await memory.getWorkingMemory({ threadId: thread.id, resourceId });
-      expect(workingMemory).not.toBeNull();
-      if (workingMemory) {
-        // Check for specific Markdown format
-        expect(workingMemory).toContain('# User Information');
-        expect(workingMemory).toContain('**First Name**: Tyler');
-        expect(workingMemory).toContain('**Location**: San Francisco');
-      }
-    });
-
-    it('should initialize with default working memory template', async () => {
-      const systemInstruction = await memory.getSystemMessage({ threadId: thread.id, resourceId });
-      expect(systemInstruction).not.toBeNull();
-      if (systemInstruction) {
-        // Should match our Markdown template
-        expect(systemInstruction).toContain('# User Information');
-        expect(systemInstruction).toContain('First Name');
-      }
-    });
-
-    it('should hide working memory tags in remembered messages', async () => {
-      const messages = [
-        createTestMessage(thread.id, 'Hi, my name is John'),
-        createTestMessage(
-          thread.id,
-          `Hello John!
-<working_memory>
-# User Information
-- **First Name**: John
-- **Last Name**: 
-- **Location**: 
-- **Interests**: 
-</working_memory>`,
-          'assistant',
-        ),
-      ];
-
-      await memory.saveMessages({ messages });
-
-      const remembered = await memory.recall({
-        threadId: thread.id,
-        perPage: 10,
-      });
-
-      // Working memory tags should be stripped from the messages
-      const content = getTextContent(remembered.messages[1]);
-      expect(content).not.toContain('<working_memory>');
-      expect(content).toContain('Hello John!');
-    });
-
-    it('should respect working memory enabled/disabled setting', async () => {
-      const dbPath = join(await mkdtemp(join(tmpdir(), `memory-working-test-${Date.now()}`)), 'test.db');
-
-      // Create memory instance with working memory disabled
-      const disabledMemory = new Memory({
-        storage: new LibSQLStore({
-          id: 'disabled-memory-storage',
-          url: `file:${dbPath}`,
-        }),
-        vector: new LibSQLVector({
-          connectionUrl: `file:${dbPath}`,
-          id: 'test-vector',
-        }),
-        embedder: openai.embedding('text-embedding-3-small'),
-        options: {
-          workingMemory: {
-            enabled: false,
-            template: `# User Information
-- **First Name**: 
-- **Last Name**:
-`,
-          },
-          lastMessages: 10,
-          semanticRecall: {
-            topK: 3,
-            messageRange: 2,
-          },
-          generateTitle: false,
-        },
-      });
-
-      const thread = await disabledMemory.saveThread({
-        thread: createTestThread('Disabled Working Memory Thread'),
-      });
-
-      const messages = [
-        createTestMessage(thread.id, 'Hi, my name is John'),
-        createTestMessage(
-          thread.id,
-          `Hello John!
-<working_memory>
-# User Information
-- **First Name**: John
-- **Last Name**: 
-</working_memory>`,
-          'assistant',
-        ),
-      ];
-
-      await disabledMemory.saveMessages({ messages });
-
-      // Working memory should be null when disabled
-      const workingMemory = await disabledMemory.getWorkingMemory({ threadId: thread.id });
-      expect(workingMemory).toBeNull();
-
-      // Thread metadata should not contain working memory
-      const updatedThread = await disabledMemory.getThreadById({ threadId: thread.id });
-      expect(updatedThread?.metadata?.workingMemory).toBeUndefined();
-    });
-
-    it('should handle LLM responses with working memory using tool calls', async () => {
-      const agent = new Agent({
-        id: 'memory-test-agent',
-        name: 'Memory Test Agent',
-        instructions: 'You are a helpful AI agent. Always add working memory tags to remember user information.',
-        model: openai('gpt-4o'),
-        memory,
-      });
-
-      const thread = await memory.createThread(createTestThread(`Tool call working memory test`));
-
-      await agent.generateLegacy('Hi, my name is Tyler and I live in San Francisco', {
-        threadId: thread.id,
-        resourceId,
-      });
-
-      const workingMemory = await memory.getWorkingMemory({ threadId: thread.id, resourceId });
-      expect(workingMemory).not.toBeNull();
-      if (workingMemory) {
-        // Check for specific Markdown format
-        expect(workingMemory).toContain('# User Information');
-        expect(workingMemory).toContain('**First Name**: Tyler');
-        expect(workingMemory).toContain('**Location**: San Francisco');
-      }
-    });
-
-    it("shouldn't pollute context with working memory tool call args, only the system instruction working memory should exist", async () => {
-      const agent = new Agent({
-        id: 'memory-test-agent',
-        name: 'Memory Test Agent',
-        instructions: 'You are a helpful AI agent. Always add working memory tags to remember user information.',
-        model: openai('gpt-4o'),
-        memory,
-      });
-
-      const thread = await memory.createThread(createTestThread(`Tool call working memory context pollution test`));
-
-      await agent.generateLegacy('Hi, my name is Tyler and I live in a submarine under the sea', {
-        threadId: thread.id,
-        resourceId,
-      });
-
-      let workingMemory = await memory.getWorkingMemory({ threadId: thread.id, resourceId });
-      expect(workingMemory).not.toBeNull();
-      if (workingMemory) {
-        expect(workingMemory).toContain('# User Information');
-        expect(workingMemory).toContain('**First Name**: Tyler');
-        expect(workingMemory?.toLowerCase()).toContain('**location**:');
-        expect(workingMemory?.toLowerCase()).toContain('submarine under the sea');
-      }
-
-      await agent.generateLegacy('I changed my name to Jim', {
-        threadId: thread.id,
-        resourceId,
-      });
-
-      workingMemory = await memory.getWorkingMemory({ threadId: thread.id, resourceId });
-      expect(workingMemory).not.toBeNull();
-      if (workingMemory) {
-        expect(workingMemory).toContain('# User Information');
-        expect(workingMemory).toContain('**First Name**: Jim');
-        expect(workingMemory?.toLowerCase()).toContain('**location**:');
-        expect(workingMemory?.toLowerCase()).toContain('submarine under the sea');
-      }
-
-      await agent.generateLegacy('I moved to Vancouver Island', {
-        threadId: thread.id,
-        resourceId,
-      });
-
-      workingMemory = await memory.getWorkingMemory({ threadId: thread.id, resourceId });
-      expect(workingMemory).not.toBeNull();
-      if (workingMemory) {
-        expect(workingMemory).toContain('# User Information');
-        expect(workingMemory).toContain('**First Name**: Jim');
-        expect(workingMemory).toContain('**Location**: Vancouver Island');
-      }
-
-      const history = await memory.recall({
-        threadId: thread.id,
-        resourceId,
-        perPage: 20,
-      });
-
-      const memoryArgs: string[] = [];
-
-      for (const message of history.messages) {
-        if (message.role === `assistant` && message.content.parts) {
-          for (const part of message.content.parts) {
-            if (typeof part === `string`) continue;
-            if (part.type === `tool-invocation` && part.toolInvocation?.toolName === `updateWorkingMemory`) {
-              memoryArgs.push((part.toolInvocation.args as any).memory);
-            }
-          }
-        }
-      }
-
-      expect(memoryArgs).not.toContain(`Tyler`);
-      expect(memoryArgs).not.toContain('submarine under the sea');
-      expect(memoryArgs).not.toContain('Jim');
-      expect(memoryArgs).not.toContain('Vancouver Island');
-      expect(memoryArgs).toEqual([]);
-
-      workingMemory = await memory.getWorkingMemory({ threadId: thread.id, resourceId });
-      expect(workingMemory).not.toBeNull();
-      if (workingMemory) {
-        // Format-specific assertion that checks for Markdown format
-        expect(workingMemory).toContain('# User Information');
-        expect(workingMemory).toContain('**First Name**: Jim');
-        expect(workingMemory).toContain('**Location**: Vancouver Island');
-      }
-    });
-
-    it('should remove tool-call/tool-result messages with toolName "updateWorkingMemory"', async () => {
-      const threadId = thread.id;
-      const messages = [
-        createTestMessage(threadId, 'User says something'),
-        // Pure tool-call message (should be removed)
-        createToolCallMessage(threadId, 'updateWorkingMemory', { key: 'value', data: 'test' }),
-        // Mixed content: tool-invocation + text (tool-invocation part should be filtered, text kept)
-
-        createToolResultMessage(threadId, 'updateWorkingMemory', randomUUID(), { memory: 'should not persist' }, [
-          { type: 'text', text: 'Normal message' },
-        ]),
-
-        // Pure text message (should be kept)
-        createTestMessage(threadId, 'Another normal message', 'assistant'),
-      ];
-
-      // Save messages
-      const { messages: saved } = await memory.saveMessages({ messages });
-
-      // Should not include any updateWorkingMemory tool-invocation messages (pure or mixed)
-      expect(
-        saved.some(
-          m =>
-            m.content.parts &&
-            m.content.parts.some(
-              p => p.type === 'tool-invocation' && p.toolInvocation?.toolName === 'updateWorkingMemory',
-            ),
-        ),
-      ).toBe(false);
-
-      // Mixed content message: should only keep the text part
-      const assistantMessages = saved.filter(m => m.role === 'assistant');
-      expect(
-        assistantMessages.every(m => {
-          return !JSON.stringify(m).includes(`updateWorkingMemory`);
-        }),
-      ).toBe(true);
-
-      // Pure text message should be present
-      expect(
-        saved.some(
-          m => m.content.parts && m.content.parts.some(p => p.type === 'text' && p.text === 'Another normal message'),
-        ),
-      ).toBe(true);
-
-      // User message should be present
-      expect(
-        saved.some(
-          m =>
-            m.content.parts && m.content.parts.some(p => p.type === 'text' && p.text?.includes('User says something')),
-        ),
-      ).toBe(true);
-    });
-  });
-
-  describe('Working Memory with agent memory', () => {
-    let agent: Agent;
-    let thread: any;
-    let memory: Memory;
-
-    beforeEach(async () => {
-      const dbPath = join(await mkdtemp(join(tmpdir(), `memory-working-test-${Date.now()}`)), 'test.db');
-      storage = new LibSQLStore({
-        id: 'test-storage',
-        url: `file:${dbPath}`,
-      });
-
-      memory = new Memory({
-        storage,
-        options: {
-          workingMemory: {
-            enabled: true,
-            schema: z.object({
-              favouriteAnimal: z.string(),
-            }),
-          },
-          lastMessages: 1,
-          generateTitle: false,
-        },
-      });
-      // Reset message counter
-      messageCounter = 0;
-
-      // Create a new thread for each test
-      thread = await memory.saveThread({
-        thread: createTestThread('Working Memory Test Thread'),
-      });
-      expect(await memory.getWorkingMemory({ threadId: thread.id, resourceId })).toBeNull();
-      agent = new Agent({
-        id: 'memory-test-agent',
-        name: 'Memory Test Agent',
-        instructions: 'You are a helpful AI agent. Always add working memory tags to remember user information.',
-        model: openai('gpt-4o'),
-        memory,
-      });
-    });
-
-    it('should remember information from working memory in subsequent calls', async () => {
-      const thread = await memory.saveThread({
-        thread: createTestThread('Remembering Test'),
-      });
-
-      // First call to establish a fact in working memory
-      await agent.generateLegacy('My favorite animal is the majestic wolf.', {
-        threadId: thread.id,
-        resourceId,
-      });
-
-      // Verify it's in the working memory
-      const workingMemoryAfterFirstCall = await memory.getWorkingMemory({ threadId: thread.id, resourceId });
-      expect(workingMemoryAfterFirstCall).not.toBeNull();
-      if (workingMemoryAfterFirstCall) {
-        expect(workingMemoryAfterFirstCall.toLowerCase()).toContain('wolf');
-      }
-
-      // add messages to the thread
-      await agent.generateLegacy('How are you doing?', {
-        threadId: thread.id,
-        resourceId,
-      });
-
-      // third call to see if the agent remembers the fact
-      const response = await agent.generateLegacy('What is my favorite animal?', {
-        threadId: thread.id,
-        resourceId,
-      });
-
-      expect(response.text.toLowerCase()).toContain('wolf');
-    });
-
-    describe('Working Memory with Schema', () => {
-      let agent: Agent;
-      beforeEach(async () => {
-        const dbPath = join(await mkdtemp(join(tmpdir(), `memory-working-test-${Date.now()}`)), 'test.db');
-        storage = new LibSQLStore({
-          id: 'test-storage',
-          url: `file:${dbPath}`,
-        });
-        vector = new LibSQLVector({
-          connectionUrl: `file:${dbPath}`,
-          id: 'test-vector',
-        });
-
-        memory = new Memory({
-          storage,
-          vector,
-          embedder: fastembed,
-          options: {
-            workingMemory: {
-              enabled: true,
-              schema: z.object({
-                city: z.string(),
-                temperature: z.number().describe('The number value of the temperature'),
-              }),
-            },
-            lastMessages: 10,
-            semanticRecall: {
-              topK: 3,
-              messageRange: 2,
-            },
-            generateTitle: false,
-          },
-        });
-        // Reset message counter
-        messageCounter = 0;
-
-        // Create a new thread for each test
-        thread = await memory.saveThread({
-          thread: createTestThread('Working Memory Test Thread'),
-        });
-
-        expect(await memory.getWorkingMemory({ threadId: thread.id, resourceId })).toBeNull();
-
-        agent = new Agent({
-          id: 'memory-test-agent',
-          name: 'Memory Test Agent',
-          instructions: `You are a helpful AI agent. Always add working memory tags to remember user information.
-          
-          Temperature, "temperature" should be reported as a number, not a string.
-          The location should be labeled "city" and reported as a string.
-          `,
-          model: openai('gpt-4o'),
-          memory,
-        });
-      });
-
-      afterEach(async () => {
-        //@ts-ignore
-        await storage.client.close();
-        //@ts-ignore
-        await vector.turso.close();
-      });
-
-      it('should accept valid working memory updates matching the schema', async () => {
-        const validMemory = { city: 'Austin', temperature: 85 };
-        await agent.generateLegacy('I am in Austin and it is 85 degrees', {
-          threadId: thread.id,
-          resourceId,
-          temperature: 0.1,
-        });
-
-        const wmRaw = await memory.getWorkingMemory({ threadId: thread.id, resourceId });
-        const wm = typeof wmRaw === 'string' ? JSON.parse(wmRaw) : wmRaw;
-        const wmObj = typeof wm === 'string' ? JSON.parse(wm) : wm;
-        expect(extractUserData(wmObj)).toMatchObject(validMemory);
-      });
-
-      it('should recall the most recent valid schema-based working memory', async () => {
-        const second = { city: 'Denver', temperature: 75 };
-        await agent.generateLegacy('Now I am in Seattle and it is 60 degrees', {
-          threadId: thread.id,
-          resourceId,
-          temperature: 0.1,
-        });
-        await agent.generateLegacy('Now I am in Denver and it is 75 degrees', {
-          threadId: thread.id,
-          resourceId,
-          temperature: 0.1,
-        });
-
-        const wmRaw = await memory.getWorkingMemory({ threadId: thread.id, resourceId });
-        const wm = typeof wmRaw === 'string' ? JSON.parse(wmRaw) : wmRaw;
-        const wmObj = typeof wm === 'string' ? JSON.parse(wm) : wm;
-        expect(extractUserData(wmObj)).toMatchObject(second);
-      });
-
-      // Skip this for now it's an edge case where an agent updates the working memory based off of the
-      // message history.
-      it.skip('should not update working from message history', async () => {
-        const newThread = await memory.saveThread({
-          thread: createTestThread('Test111'),
-        });
-        const first = { city: 'Toronto', temperature: 80 };
-        const generateOptions = {
-          memory: {
-            resource: resourceId,
-            thread: newThread.id,
-            options: {
-              lastMessages: 0,
-              semanticRecall: undefined,
-              workingMemory: {
-                enabled: true,
-                schema: z.object({
-                  city: z.string(),
-                  temperature: z.number().optional(),
-                }),
-              },
-              generateTitle: false,
-            },
-          },
-          temperature: 0,
-        } satisfies AgentGenerateOptions<any, any>;
-        await agent.generateLegacy('Now I am in Toronto and it is 80 degrees', generateOptions);
-
-        await agent.generateLegacy('how are you doing?', generateOptions);
-
-        const firstWorkingMemory = await memory.getWorkingMemory({ threadId: newThread.id, resourceId });
-        const wm = typeof firstWorkingMemory === 'string' ? JSON.parse(firstWorkingMemory) : firstWorkingMemory;
-        const wmObj = typeof wm === 'string' ? JSON.parse(wm) : wm;
-
-        expect(wmObj).toMatchObject(first);
-
-        const updatedThread = await memory.getThreadById({ threadId: newThread.id });
-        if (!updatedThread) {
-          throw new Error('Thread not found');
-        }
-        // Update thread metadata with new working memory
-        await memory.saveThread({
-          thread: {
-            ...updatedThread,
-            metadata: {
-              ...(updatedThread.metadata || {}),
-              workingMemory: { city: 'Waterloo', temperature: 78 },
-            },
-          },
-          memoryConfig: generateOptions.memory.options,
-        });
-
-        // This should not update the working memory
-        await agent.generateLegacy('how are you doing?', generateOptions);
-
-        const result = await agent.generateLegacy('Can you tell me where I am?', generateOptions);
-
-        expect(result.text).toContain('Waterloo');
-        const secondWorkingMemory = await memory.getWorkingMemory({ threadId: newThread.id, resourceId });
-        expect(secondWorkingMemory).toMatchObject({ city: 'Waterloo', temperature: 78 });
-      });
-    });
-  });
-
-  describe('Working Memory with JSONSchema7', () => {
-    let agent: Agent;
-    let thread: any;
-    let memory: Memory;
-
-    beforeEach(async () => {
-      const dbPath = join(await mkdtemp(join(tmpdir(), `memory-jsonschema-test-${Date.now()}`)), 'test.db');
-      storage = new LibSQLStore({
-        id: 'test-storage',
-        url: `file:${dbPath}`,
-      });
-      vector = new LibSQLVector({
-        connectionUrl: `file:${dbPath}`,
-        id: 'test-vector',
-      });
-
-      const jsonSchema: JSONSchema7 = {
-        type: 'object',
-        properties: {
-          name: { type: 'string' },
-          age: { type: ['number', 'null'] },
-          city: { type: 'string' },
-          preferences: {
-            type: ['object', 'null'],
-            properties: {
-              theme: { type: 'string' },
-              notifications: { type: 'boolean' },
-            },
-          },
-        },
-        required: ['name', 'city'],
-      };
-
-      memory = new Memory({
-        storage,
-        vector,
-        embedder: fastembed,
-        options: {
-          workingMemory: {
-            enabled: true,
-            schema: jsonSchema,
-          },
-          lastMessages: 10,
-          semanticRecall: {
-            topK: 3,
-            messageRange: 2,
-          },
-          generateTitle: false,
-        },
-      });
-
-      // Reset message counter
-      messageCounter = 0;
-
-      // Create a new thread for each test
-      thread = await memory.saveThread({
-        thread: createTestThread('JSONSchema7 Working Memory Test Thread'),
-      });
-
-      // Verify initial working memory is empty
-      expect(await memory.getWorkingMemory({ threadId: thread.id, resourceId })).toBeNull();
-
-      agent = new Agent({
-        id: 'jsonschema-memory-test-agent',
-        name: 'JSONSchema Memory Test Agent',
-        instructions: 'You are a helpful AI agent. Always update working memory with user information.',
-        model: openai('gpt-4o'),
-        memory,
-      });
-    });
-
-    afterEach(async () => {
-      //@ts-ignore
-      await storage.client.close();
-      //@ts-ignore
-      await vector.turso.close();
-    });
-
-    it('should accept JSONSchema7 in working memory configuration', async () => {
-      // Test that we can create a Memory instance with JSONSchema7 schema
-      const jsonSchema: JSONSchema7 = {
-        type: 'object',
-        properties: {
-          testField: { type: 'string' },
-        },
-        required: ['testField'],
-      };
-
-      const testMemory = new Memory({
-        storage,
-        options: {
-          workingMemory: {
-            enabled: true,
-            schema: jsonSchema,
-          },
-        },
-      });
-
-      // Get the working memory template
-      const template = await testMemory.getWorkingMemoryTemplate({
-        memoryConfig: {
-          workingMemory: {
-            enabled: true,
-            schema: jsonSchema,
-          },
-        },
-      });
-
-      expect(template).not.toBeNull();
-      expect(template?.format).toBe('json');
-      expect(template?.content).toContain('testField');
-      expect(template?.content).toContain('string');
-    });
-
-    it('should accept valid working memory updates matching the JSONSchema7', async () => {
-      await agent.generateLegacy(
-        'Hi, my name is John Doe, I am 30 years old and I live in Boston. I prefer dark theme and want notifications enabled.',
-        {
-          threadId: thread.id,
-          resourceId,
-        },
-      );
-
-      const wmRaw = await memory.getWorkingMemory({ threadId: thread.id, resourceId });
-      const wm = typeof wmRaw === 'string' ? JSON.parse(wmRaw) : wmRaw;
-      const wmObj = typeof wm === 'string' ? JSON.parse(wm) : wm;
-      const userData = extractUserData(wmObj);
-
-      expect(userData.name).toBe('John Doe');
-      expect(userData.age).toBe(30);
-      expect(userData.city).toBe('Boston');
-    });
-
-    it('should handle required and optional fields correctly with JSONSchema7', async () => {
-      // Test with only required fields
-      await agent.generateLegacy('My name is Jane Smith and I live in Portland.', {
-        threadId: thread.id,
-        resourceId,
-      });
-
-      const wmRaw = await memory.getWorkingMemory({ threadId: thread.id, resourceId });
-      const wm = typeof wmRaw === 'string' ? JSON.parse(wmRaw) : wmRaw;
-      const wmObj = typeof wm === 'string' ? JSON.parse(wm) : wm;
-      const userData = extractUserData(wmObj);
-
-      expect(userData.name).toBe('Jane Smith');
-      expect(userData.city).toBe('Portland');
-      // Age is not required, so it might not be set
-    });
-
-    it('should update working memory progressively with JSONSchema7', async () => {
-      // First message with partial info
-      await agent.generateLegacy('Hi, I am Alex and I live in Miami.', {
-        threadId: thread.id,
-        resourceId,
-      });
-
-      let wmRaw = await memory.getWorkingMemory({ threadId: thread.id, resourceId });
-      let wm = typeof wmRaw === 'string' ? JSON.parse(wmRaw) : wmRaw;
-      let wmObj = typeof wm === 'string' ? JSON.parse(wm) : wm;
-      let userData = extractUserData(wmObj);
-
-      expect(userData.name).toBe('Alex');
-      expect(userData.city).toBe('Miami');
-
-      // Second message adding more info
-      await agent.generateLegacy('I am 25 years old.', {
-        threadId: thread.id,
-        resourceId,
-      });
-
-      wmRaw = await memory.getWorkingMemory({ threadId: thread.id, resourceId });
-      wm = typeof wmRaw === 'string' ? JSON.parse(wmRaw) : wmRaw;
-      wmObj = typeof wm === 'string' ? JSON.parse(wm) : wm;
-      userData = extractUserData(wmObj);
-
-      expect(userData.name).toBe('Alex');
-      expect(userData.city).toBe('Miami');
-      expect(userData.age).toBe(25);
-    });
-
-    it('should persist working memory across multiple interactions with JSONSchema7', async () => {
-      // Set initial data
-      await agent.generateLegacy('My name is Sarah Wilson, I am 28 and live in Seattle.', {
-        threadId: thread.id,
-        resourceId,
-      });
-
-      // Verify working memory is set
-      let wmRaw = await memory.getWorkingMemory({ threadId: thread.id, resourceId });
-      let wm = typeof wmRaw === 'string' ? JSON.parse(wmRaw) : wmRaw;
-      let wmObj = typeof wm === 'string' ? JSON.parse(wm) : wm;
-      let userData = extractUserData(wmObj);
-      expect(userData.name).toBe('Sarah Wilson');
-
-      // Ask a question that should use the working memory
-      const response = await agent.generateLegacy('What is my name and where do I live?', {
-        threadId: thread.id,
-        resourceId,
-      });
-
-      // The response should contain the information from working memory
-      expect(response.text.toLowerCase()).toContain('sarah');
-      expect(response.text.toLowerCase()).toContain('seattle');
-    });
-  });
-
-  describe('Resource-Scoped Working Memory Tests', () => {
-    beforeEach(async () => {
-      // Create a new unique database file in the temp directory for each test
-      const dbPath = join(await mkdtemp(join(tmpdir(), `memory-resource-working-test-`)), 'test.db');
-      console.log('dbPath', dbPath);
-
-      storage = new LibSQLStore({
-        id: 'test-storage',
-        url: `file:${dbPath}`,
-      });
-      vector = new LibSQLVector({
-        connectionUrl: `file:${dbPath}`,
-        id: 'test-vector',
-      });
-
-      // Create memory instance with resource-scoped working memory enabled
-      memory = new Memory({
-        options: {
-          workingMemory: {
-            enabled: true,
-            scope: 'resource',
-            template: `# User Information
-- **First Name**: 
-- **Last Name**: 
-- **Location**: 
-- **Interests**: 
-`,
-          },
-          lastMessages: 10,
-          semanticRecall: {
-            topK: 3,
-            messageRange: 2,
-          },
-          generateTitle: false,
-        },
-        storage,
-        vector,
-        embedder: fastembed,
-      });
-      // Reset message counter
-      messageCounter = 0;
-      // Create a new thread for each test
-      thread = await memory.saveThread({
-        thread: createTestThread('Resource Working Memory Test Thread'),
-      });
-    });
-
-    afterEach(async () => {
-      //@ts-ignore
-      await storage.client.close();
-      //@ts-ignore
-      await vector.turso.close();
-    });
-
-    it('should store working memory at resource level', async () => {
-      // Update working memory using the updateWorkingMemory method
-      const workingMemoryData = `# User Information
-- **First Name**: John
-- **Last Name**: Doe
-- **Location**: New York
-- **Interests**: AI, Machine Learning
-`;
-
-      await memory.updateWorkingMemory({
-        threadId: thread.id,
-        resourceId,
-        workingMemory: workingMemoryData,
-      });
-
-      // Get working memory and verify it's stored at resource level
-      const retrievedWorkingMemory = await memory.getWorkingMemory({
-        threadId: thread.id,
-        resourceId,
-      });
-
-      expect(retrievedWorkingMemory).toBe(workingMemoryData);
-    });
-
-    it('should share working memory across multiple threads for the same resource', async () => {
-      // Create a second thread for the same resource
-      const thread2 = await memory.saveThread({
-        thread: createTestThread('Second Resource Working Memory Test Thread'),
-      });
-
-      // Update working memory from first thread
-      const workingMemoryData = `# User Information
-- **First Name**: Alice
-- **Last Name**: Smith
-- **Location**: California
-- **Interests**: Data Science, Python
-`;
-
-      await memory.updateWorkingMemory({
-        threadId: thread.id,
-        resourceId,
-        workingMemory: workingMemoryData,
-      });
-
-      // Retrieve working memory from second thread
-      const retrievedFromThread2 = await memory.getWorkingMemory({
-        threadId: thread2.id,
-        resourceId,
-      });
-
-      expect(retrievedFromThread2).toBe(workingMemoryData);
-    });
-
-    it('should update working memory across all threads when updated from any thread', async () => {
-      // Create multiple threads for the same resource
-      const thread2 = await memory.saveThread({
-        thread: createTestThread('Second Thread'),
-      });
-      const thread3 = await memory.saveThread({
-        thread: createTestThread('Third Thread'),
-      });
-
-      // Set initial working memory from thread1
-      const initialWorkingMemory = `# User Information
-- **First Name**: Bob
-- **Last Name**: Johnson
-- **Location**: Texas
-- **Interests**: Software Development
-`;
-
-      await memory.updateWorkingMemory({
-        threadId: thread.id,
-        resourceId,
-        workingMemory: initialWorkingMemory,
-      });
-
-      // Update working memory from thread2
-      const updatedWorkingMemory = `# User Information
-- **First Name**: Bob
-- **Last Name**: Johnson
-- **Location**: Florida
-- **Interests**: Software Development, Travel
-`;
-
-      await memory.updateWorkingMemory({
-        threadId: thread2.id,
-        resourceId,
-        workingMemory: updatedWorkingMemory,
-      });
-
-      // Verify all threads see the updated working memory
-      const wmFromThread1 = await memory.getWorkingMemory({ threadId: thread.id, resourceId });
-      const wmFromThread2 = await memory.getWorkingMemory({ threadId: thread2.id, resourceId });
-      const wmFromThread3 = await memory.getWorkingMemory({ threadId: thread3.id, resourceId });
-
-      expect(wmFromThread1).toBe(updatedWorkingMemory);
-      expect(wmFromThread2).toBe(updatedWorkingMemory);
-      expect(wmFromThread3).toBe(updatedWorkingMemory);
-    });
-
-    it('should handle JSON format correctly for resource-scoped working memory', async () => {
-      const workingMemoryData = `{"name":"Charlie","age":30,"city":"Seattle"}`;
-
-      await memory.updateWorkingMemory({
-        threadId: thread.id,
-        resourceId,
-        workingMemory: workingMemoryData,
-      });
-
-      // Test JSON format retrieval
-      const retrievedAsJson = await memory.getWorkingMemory({
-        threadId: thread.id,
-        resourceId,
-      });
-
-      expect(retrievedAsJson).toBe(`{"name":"Charlie","age":30,"city":"Seattle"}`);
-
-      // Test default format retrieval
-      const retrievedDefault = await memory.getWorkingMemory({
-        threadId: thread.id,
-        resourceId,
-      });
-
-      expect(retrievedDefault).toBe(workingMemoryData);
-    });
-
-    it('should verify storage adapter support for resource working memory', async () => {
-      // This test would require a mock storage adapter that doesn't support resource working memory
-      // For now, we'll just verify that LibSQL supports it
-      expect(storage.supports.resourceWorkingMemory).toBe(true);
-    });
-
-    it('should initialize working memory when creating new threads for existing resources', async () => {
-      // Create first thread and set working memory
-      const thread1 = await memory.saveThread({
-        thread: createTestThread('First Thread'),
-      });
-
-      const workingMemoryData = `# User Information
-- **First Name**: David
-- **Last Name**: Wilson
-- **Location**: Portland
-- **Interests**: Music, Photography
-`;
-
-      await memory.updateWorkingMemory({
-        threadId: thread1.id,
-        resourceId,
-        workingMemory: workingMemoryData,
-      });
-
-      // Create a new thread for the same resource
-      const thread2 = await memory.saveThread({
-        thread: createTestThread('Second Thread'),
-      });
-
-      // The new thread should immediately have access to the existing working memory
-      const retrievedMemory = await memory.getWorkingMemory({
-        threadId: thread2.id,
-        resourceId,
-      });
-
-      expect(retrievedMemory).toBe(workingMemoryData);
-    });
-
-    describe('Setting Working Memory via Thread Metadata (Resource Scope)', () => {
-      it('should set working memory at resource level when creating thread with metadata.workingMemory', async () => {
-        const workingMemoryData = `# User Information
-- **First Name**: John
-- **Last Name**: Doe
-- **Location**: New York`;
-
-        const thread = await memory.createThread({
-          resourceId,
-          metadata: {
-            workingMemory: workingMemoryData,
-          },
-        });
-
-        const retrievedWM = await memory.getWorkingMemory({
-          threadId: thread.id,
-          resourceId,
-        });
-
-        expect(retrievedWM).toBe(workingMemoryData);
-      });
-
-      it('should store working memory in resource table, not thread metadata', async () => {
-        const workingMemoryData = `# User Info
-- Name: Alice
-- Email: alice@example.com`;
-
-        const thread = await memory.createThread({
-          resourceId,
-          metadata: {
-            workingMemory: workingMemoryData,
-          },
-        });
-
-        // Verify it's in the resource table
-        const resource = await storage.getResourceById({ resourceId });
-        expect(resource?.workingMemory).toBe(workingMemoryData);
-
-        // The working memory should come from resource, not thread metadata
-        const wmFromResource = await memory.getWorkingMemory({
-          threadId: thread.id,
-          resourceId,
-        });
-        expect(wmFromResource).toBe(workingMemoryData);
-      });
-
-      it('should share working memory across threads when set via metadata on first thread', async () => {
-        const workingMemoryData = `# Shared User Data
-- Name: Bob
-- Preferences: Dark Mode`;
-
-        // Create first thread with working memory
-        const thread1 = await memory.createThread({
-          resourceId,
-          metadata: {
-            workingMemory: workingMemoryData,
-          },
-        });
-
-        // Create second thread without working memory
-        const thread2 = await memory.createThread({
-          resourceId,
-        });
-
-        // Both threads should see the same working memory
-        const wm1 = await memory.getWorkingMemory({ threadId: thread1.id, resourceId });
-        const wm2 = await memory.getWorkingMemory({ threadId: thread2.id, resourceId });
-
-        expect(wm1).toBe(workingMemoryData);
-        expect(wm2).toBe(workingMemoryData);
-      });
-
-      it('should produce same result as updateWorkingMemory for resource scope', async () => {
-        const workingMemoryData = `# User Profile
-- ID: user-123
-- Subscription: Premium`;
-
-        // Method 1: Via metadata (this should work but doesn't)
-        const thread1 = await memory.createThread({
-          resourceId: 'resource-1',
-          metadata: { workingMemory: workingMemoryData },
-        });
-
-        // Method 2: Via updateWorkingMemory (this works)
-        const thread2 = await memory.createThread({ resourceId: 'resource-2' });
-        await memory.updateWorkingMemory({
-          threadId: thread2.id,
-          resourceId: 'resource-2',
-          workingMemory: workingMemoryData,
-        });
-
-        const wm1 = await memory.getWorkingMemory({
-          threadId: thread1.id,
-          resourceId: 'resource-1',
-        });
-        const wm2 = await memory.getWorkingMemory({
-          threadId: thread2.id,
-          resourceId: 'resource-2',
-        });
-
-        expect(wm1).toBe(wm2);
-        expect(wm1).toBe(workingMemoryData);
-      });
-
-      it('should update working memory at resource level when updating thread metadata', async () => {
-        const initialWM = `# User Data
-- Name: Charlie`;
-
-        const thread = await memory.createThread({
-          resourceId,
-          metadata: { workingMemory: initialWM },
-        });
-
-        // Verify initial working memory is set
-        const retrievedInitial = await memory.getWorkingMemory({
-          threadId: thread.id,
-          resourceId,
-        });
-        expect(retrievedInitial).toBe(initialWM);
-
-        // Update thread metadata with new working memory
-        const updatedWM = `# User Data
-- Name: Charlie
-- Email: charlie@example.com`;
-
-        await memory.updateThread({
-          id: thread.id,
-          title: thread.title || 'Test Thread',
-          metadata: {
-            ...thread.metadata,
-            workingMemory: updatedWM,
-          },
-        });
-
-        // Verify working memory was updated in resource table
-        const retrievedUpdated = await memory.getWorkingMemory({
-          threadId: thread.id,
-          resourceId,
-        });
-        expect(retrievedUpdated).toBe(updatedWM);
-      });
-    });
-  });
-
-  describe('Setting Working Memory via Thread Metadata (Thread Scope)', () => {
-    let memory: Memory;
-    let storage: LibSQLStore;
-    let vector: LibSQLVector;
-
-    beforeEach(async () => {
-      const dbPath = join(await mkdtemp(join(tmpdir(), `memory-thread-working-test-`)), 'test.db');
-
-      storage = new LibSQLStore({
-        id: 'test-storage',
-        url: `file:${dbPath}`,
-      });
-      vector = new LibSQLVector({
-        connectionUrl: `file:${dbPath}`,
-        id: 'test-vector',
-      });
-
-      // Create memory instance with thread-scoped working memory (explicitly set)
-      memory = new Memory({
-        options: {
-          workingMemory: {
-            enabled: true,
-            scope: 'thread', // Explicitly set to thread-scoped (default is now 'resource')
-            template: `# User Information
-- **First Name**: 
-- **Last Name**: 
-- **Location**: 
-`,
-          },
-          lastMessages: 10,
-          semanticRecall: {
-            topK: 3,
-            messageRange: 2,
-          },
-          generateTitle: false,
-        },
-        storage,
-        vector,
-        embedder: fastembed,
-      });
-    });
-
-    afterEach(async () => {
-      //@ts-ignore
-      await storage.client.close();
-      //@ts-ignore
-      await vector.turso.close();
-    });
-
-    it('should set working memory when creating thread with metadata.workingMemory', async () => {
-      const workingMemoryData = `# User Information
-- **First Name**: Jane
-- **Last Name**: Smith
-- **Location**: Boston`;
-
-      const thread = await memory.createThread({
-        resourceId,
-        metadata: {
-          workingMemory: workingMemoryData,
-        },
-      });
-
-      const retrievedWM = await memory.getWorkingMemory({ threadId: thread.id });
-      expect(retrievedWM).toBe(workingMemoryData);
-    });
-
-    it('should update working memory when updating thread metadata', async () => {
-      const initialWM = `# Patient Profile
-- Name: John Doe`;
-
-      const thread = await memory.createThread({
-        resourceId,
-        metadata: { workingMemory: initialWM },
-      });
-
-      const updatedWM = `# Patient Profile
-- Name: John Doe
-- Blood Type: O+`;
-
-      await memory.updateThread({
-        id: thread.id,
-        title: thread.title || 'Test Thread',
-        metadata: {
-          ...thread.metadata,
-          workingMemory: updatedWM,
-        },
-      });
-
-      const retrievedWM = await memory.getWorkingMemory({ threadId: thread.id });
-      expect(retrievedWM).toBe(updatedWM);
-    });
-
-    it('should produce same result as updateWorkingMemory for thread scope', async () => {
-      const workingMemoryData = `# User Info
-- Name: Jane Smith
-- Location: NYC`;
-
-      // Method 1: Via metadata
-      const thread1 = await memory.createThread({
-        resourceId,
-        metadata: { workingMemory: workingMemoryData },
-      });
-
-      // Method 2: Via updateWorkingMemory
-      const thread2 = await memory.createThread({ resourceId });
-      await memory.updateWorkingMemory({
-        threadId: thread2.id,
-        resourceId,
-        workingMemory: workingMemoryData,
-      });
-
-      const wm1 = await memory.getWorkingMemory({ threadId: thread1.id });
-      const wm2 = await memory.getWorkingMemory({ threadId: thread2.id });
-
-      expect(wm1).toBe(wm2);
-      expect(wm1).toBe(workingMemoryData);
-    });
-
-    it('should work as documented in medical consultation example', async () => {
-      // Exact example from docs
-      const thread = await memory.createThread({
-        threadId: 'thread-123',
-        resourceId: 'user-456',
-        title: 'Medical Consultation',
-        metadata: {
-          workingMemory: `# Patient Profile
-- Name: John Doe
-- Blood Type: O+
-- Allergies: Penicillin
-- Current Medications: None
-- Medical History: Hypertension (controlled)`,
-        },
-      });
-
-      const wm = await memory.getWorkingMemory({ threadId: thread.id });
-      expect(wm).toContain('John Doe');
-      expect(wm).toContain('O+');
-      expect(wm).toContain('Penicillin');
-    });
-  });
+// v4
+getWorkingMemoryTests({
+  model: openai('gpt-4o'),
+});
+
+// v5
+getWorkingMemoryTests({
+  model: 'openai/gpt-4o',
+});
+
+// v6
+getWorkingMemoryTests({
+  model: openaiV6('gpt-4o'),
 });
